backend/requirements.txt
fastapi==0.121.0
uvicorn
pydantic==2.12.4
python-dotenv
firebase-admin==7.1.0
